balance/transaction calculations with the state

+++ migrations/0002_create_transactions_table.up.sql
@@ -5,7 +5,7 @@ CREATE TABLE transactions
     transaction_id VARCHAR(255) NOT NULL UNIQUE, --unique creates a unique index(?) for faster search.
     --if not, maybe a cache system would do better
     state VARCHAR(15) NOT NULL,
-    amount BIGINT NOT NULL CHECK (amount >= 0),
+    amount NUMERIC(20,2) NOT NULL CHECK (amount >= 0),
 
     created_at     TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
